- Reorganizing code for FamiliesService/family.js/dbmodels.js
