+++ utils/family.js
@@ -3,59 +3,30 @@
  *
  */
 const Sequelize = require("sequelize");
-const conn = require("../databases.js").getConn_Dfam();
 
-const familyModel = require("../models/family.js")(conn, Sequelize);
-const aliasModel = require("../models/family_database_alias.js")(conn, Sequelize);
-const classificationModel = require("../models/classification.js")(conn, Sequelize);
-const rmTypeModel = require("../models/repeatmasker_type.js")(conn, Sequelize);
-const rmSubTypeModel = require("../models/repeatmasker_subtype.js")(conn, Sequelize);
-const rmStageModel = require("../models/repeatmasker_stage.js")(conn, Sequelize);
-const familyHasBufferStageModel = require("../models/family_has_buffer_stage.js")(conn, Sequelize);
-const familyHasCitationModel = require("../models/family_has_citation.js")(conn, Sequelize);
-const citationModel = require("../models/citation.js")(conn, Sequelize);
-const dfamTaxdbModel = require("../models/dfam_taxdb.js")(conn, Sequelize);
-const curationStateModel = require("../models/curation_state.js")(conn, Sequelize);
-const familyAssemblyDataModel = require("../models/family_assembly_data.js")(conn, Sequelize);
-const assemblyModel = require("../models/assembly.js")(conn, Sequelize);
-const codingSequenceModel = require("../models/coding_sequence.js")(conn, Sequelize);
-
-familyModel.hasMany(aliasModel, { foreignKey: 'family_id', as: 'aliases' });
-familyModel.belongsTo(curationStateModel, { foreignKey: 'curation_state_id', as: 'curation_state' });
-familyModel.belongsTo(classificationModel, { foreignKey: 'classification_id', as: 'classification' });
-familyModel.belongsToMany(rmStageModel, { as: 'search_stages', through: 'family_has_search_stage', foreignKey: 'family_id', otherKey: 'repeatmasker_stage_id' });
-familyModel.belongsToMany(rmStageModel, { as: 'buffer_stages', through: familyHasBufferStageModel, foreignKey: 'family_id', otherKey: 'repeatmasker_stage_id' });
-familyModel.belongsToMany(citationModel, { as: 'citations', through: familyHasCitationModel, foreignKey: 'family_id', otherKey: 'citation_pmid' });
-familyModel.belongsToMany(dfamTaxdbModel, { as: 'clades', through: 'family_clade', foreignKey: 'family_id', otherKey: 'dfam_taxdb_tax_id' });
-familyModel.hasMany(familyAssemblyDataModel, { as: 'family_assembly_data', foreignKey: 'family_id' });
-familyModel.hasMany(codingSequenceModel, { foreignKey: 'family_id', as: 'coding_sequences' });
-
-familyAssemblyDataModel.belongsTo(assemblyModel, { foreignKey: 'assembly_id' });
-
-assemblyModel.belongsTo(dfamTaxdbModel, { foreignKey: 'dfam_taxdb_tax_id' });
-
-classificationModel.belongsTo(rmTypeModel, { foreignKey: 'repeatmasker_type_id', as: 'rm_type' });
-classificationModel.belongsTo(rmSubTypeModel, { foreignKey: 'repeatmasker_subtype_id', as: 'rm_subtype' });
+const dfam = require("../databases").getModels_Dfam();
+const dfam_user = require("../databases").getModels_User();
+const mapFields = require("../utils/mapFields");
 
 function getFamilyForAnnotation(accession) {
-  return familyModel.findOne({
+  return dfam.familyModel.findOne({
     attributes: [ "id", "name", "accession", "version", "length", "title", "description", "author", "refineable", "consensus", "hmm_general_threshold" ],
     where: { accession },
     include: [
       'aliases',
-      { model: classificationModel, as: 'classification', include: [ 'rm_type', 'rm_subtype' ] },
+      { model: dfam.classificationModel, as: 'classification', include: [ 'rm_type', 'rm_subtype' ] },
       'curation_state',
-      { model: rmStageModel, as: 'search_stages' },
-      { model: rmStageModel, as: 'buffer_stages', through: {
+      { model: dfam.rmStageModel, as: 'search_stages' },
+      { model: dfam.rmStageModel, as: 'buffer_stages', through: {
         attributes: [ 'start_pos', 'end_pos'],
       } },
-      { model: citationModel, as: 'citations', through: { attributes: [ 'order_added', 'comment' ] } },
+      { model: dfam.citationModel, as: 'citations', through: { attributes: [ 'order_added', 'comment' ] } },
       'clades',
-      { model: familyAssemblyDataModel, as: 'family_assembly_data',
+      { model: dfam.familyAssemblyDataModel, as: 'family_assembly_data',
         attributes: ['hmm_hit_GA', 'hmm_hit_TC', 'hmm_hit_NC', 'hmm_fdr'],
         include: [
-          { model: assemblyModel, include: [
-            { model: dfamTaxdbModel, attributes: ['tax_id', 'sanitized_name', 'scientific_name'] }
+          { model: dfam.assemblyModel, include: [
+            { model: dfam.dfamTaxdbModel, attributes: ['tax_id', 'sanitized_name', 'scientific_name'] }
           ] },
         ],
       },
@@ -82,7 +53,7 @@ function getFamilyForAnnotation(accession) {
 };
 
 function getFamilyWithConsensus(accession) {
-  return familyModel.findOne({
+  return dfam.familyModel.findOne({
     attributes: [ "id", "name", "accession", "version", "consensus" ],
     where: { accession },
   }).then(function(family) {
@@ -94,8 +65,343 @@ function getFamilyWithConsensus(accession) {
   });
 };
 
+function familySubqueries(rows, format) {
+  return Promise.all(rows.map(function(row) {
+    const family_id = row.id;
+
+    const subqueries = [];
+
+    subqueries.push(dfam.familyCladeModel.findAll({
+      where: { family_id },
+      include: { model: dfam.dfamTaxdbModel, as: 'dfam_taxdb', attributes: ["lineage"], }
+    }).then(fcs => row.clades = fcs.map(fc => {
+      return { lineage: fc.dfam_taxdb.lineage };
+    })));
+
+    if (format == "full") {
+      subqueries.push(dfam.familyDatabaseAliasModel.findAll({
+        attributes: ["db_id", "db_link"],
+        where: { family_id },
+      }).then(as => row.aliases = as));
+
+      subqueries.push(dfam.familyHasSearchStageModel.findAll({
+        where: { family_id },
+        include: { model: dfam.rmStageModel, as: 'repeatmasker_stage', attributes: ["name"] }
+      }).then(fss => row.search_stages = fss.map(fs => {
+        return { name: fs.repeatmasker_stage.name };
+      })));
+
+      subqueries.push(dfam.familyHasBufferStageModel.findAll({
+        attributes: ["start_pos", "end_pos"],
+        where: { family_id },
+        include: { model: dfam.rmStageModel, as: 'repeatmasker_stage', attributes: ["name"] }
+      }).then(fbs => row.buffer_stages = fbs.map(fs => {
+        return { name: fs.repeatmasker_stage.name, family_has_buffer_stage: {
+          start_pos: fs.start_pos,
+          end_pos: fs.end_pos,
+        } };
+      })));
+
+      subqueries.push(dfam.familyHasCitationModel.findAll({
+        where: { family_id },
+        include: { model: dfam.citationModel, as: 'citation', attributes: [
+          "pmid", "title", "authors", "journal", "pubdate",
+        ] },
+        order: [ ['order_added', 'ASC'] ],
+      }).then(fhcs => row.citations = fhcs.map(fhc => {
+        return {
+          pmid: fhc.citation.pmid,
+          title: fhc.citation.title,
+          authors: fhc.citation.authors,
+          journal: fhc.citation.journal,
+          pubdate: fhc.citation.pubdate,
+        };
+      })));
+
+      subqueries.push(dfam_user.userModel.findOne({ where: { id: row.deposited_by_id }, attributes: [ 'full_name' ] }).then(function(user) {
+        if (user) {
+          row.submitter = user.full_name;
+        }
+      }));
+
+      subqueries.push(dfam.familyFeatureModel.findAll({
+        where: { family_id: row.id },
+        include: [
+          { model: dfam.featureAttributeModel, as: 'feature_attributes' }
+        ] }).then(function(features) { row.features = features; }));
+      subqueries.push(dfam.codingSequenceModel.findAll({
+        where: { family_id: row.id }
+      }).then(function(coding_sequences) { row.coding_sequences = coding_sequences; }));
+    }
+
+    return Promise.all(subqueries).then(() => row);
+  }));
+}
+
+
+function familyQueryRowToObject(row, format) {
+  const obj = mapFields(row, {}, {
+    "accession": "accession",
+    "name": "name",
+    "version": "version",
+    "title": "title",
+    "description": "description",
+    "length": "length",
+  });
+
+  if (row.classification) {
+    obj.classification = row.classification.lineage;
+    if (row.classification.rm_type) {
+      obj.repeat_type_name = row.classification.rm_type.name;
+    }
+    if (row.classification.rm_subtype) {
+      obj.repeat_subtype_name = row.classification.rm_subtype.name;
+    }
+  }
+
+  obj["clades"] = [];
+  if (row.clades) {
+    obj["clades"] = row.clades.map(cl => cl.lineage);
+  }
+
+  if (format == "summary") {
+    return obj;
+  }
+
+  mapFields(row, obj, {
+    "consensus": "consensus_sequence",
+    "author": "author",
+    "submitter": "submitter",
+    "date_created": "date_created",
+    "date_modified": "date_modified",
+    "target_site_cons": "target_site_cons",
+    "refineable": "refineable",
+    "disabled": "disabled",
+    "model_mask": "model_mask",
+    "hmm_general_threshold": "hmm_general_threshold",
+    "source_method_desc": "source_method_description",
+  });
+
+  if (obj.refineable !== undefined) {
+    obj.refineable = Boolean(obj.refineable);
+  }
+
+  if (obj.disabled !== undefined) {
+    obj.disabled = Boolean(obj.disabled);
+  }
+
+  if (row.curation_state) {
+    obj.curation_state_name = row.curation_state.name;
+    obj.curation_state_description = row.curation_state.description;
+  }
+
+  if (row.source_method) {
+    obj.source_method = row.source_method.name;
+  }
+
+  if (row.source_assembly) {
+    obj.source_assembly = {
+      label: `${row.source_assembly.name}: ${row.source_assembly.description}`,
+      hyperlink: row.source_assembly.uri,
+    };
+  }
+
+  const aliases = obj["aliases"] = [];
+  if (row.aliases) {
+    row.aliases.forEach(function(alias) {
+      aliases.push(mapFields(alias, {}, { "db_id": "database", "db_link": "alias" }));
+    });
+  }
+
+  const search_stages = obj["search_stages"] = [];
+  if (row.search_stages) {
+    row.search_stages.forEach(function(ss) {
+      search_stages.push({ name: ss.name });
+    });
+  }
+
+  const buffer_stages = obj["buffer_stages"] = [];
+  if (row.buffer_stages) {
+    row.buffer_stages.forEach(function(bs) {
+      buffer_stages.push({
+        name: bs.name,
+        start: bs.family_has_buffer_stage.start_pos,
+        end: bs.family_has_buffer_stage.end_pos,
+      });
+    });
+  }
+
+  const citations = obj["citations"] = [];
+  if (row.citations) {
+    row.citations.forEach(function(c) {
+      citations.push(mapFields(c, {}, {
+        "pmid": "pmid",
+        "title": "title",
+        "authors": "authors",
+        "journal": "journal",
+        "pubdate": "pubdate",
+      }));
+    });
+  }
+
+  const features = obj["features"] = [];
+  if (row.features) {
+    row.features.forEach(function(f) {
+      const feature = mapFields(f, {}, {
+        "feature_type": "type",
+        "description": "description",
+        "model_start_pos": "model_start_pos",
+        "model_end_pos": "model_end_pos",
+        "label": "label",
+      });
+      feature.attributes = [];
+      f.feature_attributes.forEach(function(a) {
+        feature.attributes.push(mapFields(a, {}, {
+          "attribute": "attribute",
+          "value": "value",
+        }));
+      });
+      features.push(feature);
+    });
+  }
+
+  const coding_seqs = obj["coding_seqs"] = [];
+  if (row.coding_sequences) {
+    row.coding_sequences.forEach(function(cs) {
+      const cds = mapFields(cs, {}, {
+        "product": "product",
+        "translation": "translation",
+        "cds_start": "start",
+        "cds_end": "end",
+        "exon_count": "exon_count",
+        "external_reference": "external_reference",
+        "reverse": "reverse",
+        "stop_codons": "stop_codons",
+        "frameshifts": "frameshifts",
+        "gaps": "gaps",
+        "percent_identity": "percent_identity",
+        "left_unaligned": "left_unaligned",
+        "right_unaligned": "right_unaligned",
+        "classification_id": "classification_id",
+        "align_data": "align_data",
+        "description": "description",
+        "protein_type": "protein_type",
+      });
+
+      // 'reverse' is stored as an integer
+      cds.reverse = !!cds.reverse;
+
+      // Coordinates are stored 0-based half open, but the API needs to return
+      // 1-based closed. To accomplish this, 1 is added to start and exon_starts.
+      // TODO: When the database becomes 1-based closed, remove the adjusting code.
+      if (cds.start !== undefined && cds.start !== null) {
+        cds.start += 1;
+      }
+
+      if (cs.exon_starts) {
+        cds.exon_starts = cs.exon_starts.toString().split(",").map(x => parseInt(x) + 1);
+      }
+      if (cs.exon_ends) {
+        cds.exon_ends = cs.exon_ends.toString().split(",").map(x => parseInt(x));
+      }
+
+      coding_seqs.push(cds);
+    });
+  }
+
+  return obj;
+}
+
+// Helper function for collecting ancestor/descendant clade information
+// Returns a promise.
+// If the specified clade is not present, the result is null.
+// Otherwise, result.ids is a list of IDs (self + ancestors) and
+// result.lineage is a lineage string (useful for searching descendants).
+// result.ids will contain ancestors only if clade_relatives is "ancestors" or "both"
+async function collectClades(clade, clade_relatives) {
+  if (!clade) {
+    return null;
+  }
+
+  const result = { ids: [], lineage: null };
+
+
+  // Try clade by ID first
+  const id = parseInt(clade);
+  let record;
+  if (!isNaN(id)) {
+    record = await dfam.ncbiTaxdbNamesModel.findOne({
+      where: { tax_id: id, name_class: 'scientific name' },
+      attributes: [ 'tax_id', 'name_txt' ],
+      include: [
+        { model: dfam.ncbiTaxdbNodesModel, attributes: [ "parent_id" ] }
+      ]
+    });
+  }
+
+  // Then try by scientific name
+  if (!record) {
+    record = await dfam.ncbiTaxdbNamesModel.findOne({
+      where: { name_class: 'scientific name', name_txt: clade },
+      attributes: [ 'tax_id', 'name_txt' ],
+      include: [
+        { model: dfam.ncbiTaxdbNodesModel, attributes: [ "parent_id" ] },
+      ],
+    });
+  }
+
+  if (!record) {
+    return null;
+  }
+
+  // Primary results: the given ID and its lineage
+  result.ids.push(record.tax_id);
+  result.lineage = record.name_txt;
+
+  // Secondary query: parent IDs
+  const recurseParents = async function(parent_id) {
+    const parent = await dfam.ncbiTaxdbNodesModel.findOne({
+      attributes: [ 'tax_id', 'parent_id' ],
+      where: { tax_id: parent_id },
+      include: [
+        { model: dfam.ncbiTaxdbNamesModel, where: { name_class: 'scientific name' }, attributes: [ 'name_txt' ] },
+      ]
+    });
+
+    if (parent) {
+      if (clade_relatives === "ancestors" || clade_relatives === "both") {
+        result.ids.push(parent.tax_id);
+      }
+
+      let parent_name = "";
+      if (parent.ncbi_taxdb_names.length) {
+        parent_name = parent.ncbi_taxdb_names[0].name_txt;
+      }
+      result.lineage = parent_name + ";" + result.lineage;
+
+      if (parent_id !== 1) {
+        return recurseParents(parent.parent_id);
+      }
+    }
+  };
+
+  let recurseParentsPromise;
+  if (record.tax_id !== 1) {
+    recurseParentsPromise = recurseParents(record.ncbi_taxdb_node.parent_id);
+  } else {
+    recurseParentsPromise = Promise.resolve();
+  }
+
+  await recurseParentsPromise;
+  return result;
+}
+
+
 
 module.exports = {
   getFamilyForAnnotation,
-  getFamilyWithConsensus
+  getFamilyWithConsensus,
+  familySubqueries,
+  familyQueryRowToObject,
+  collectClades
 };
